Rollback previous changes
The plan with the userReport page was to bring more data from the
userLogs to the same report, but it proved stupidly difficult via the
Parse API, so it has been moved out to the R analytics scripts. Hence
rolling back this page to just pull from the user table. It now doesn’t
add much against the activity dashboard, except massively reduced load
time